Added 'suffixes' argument to merge.data.table function.
The defaults are currently set to c("", ".1") so that the function behaves like before. The user can (of course) set this to whatever they like.

I made a note in the documentation about this argument and mentioned that the default will change to c(".x", ".y") to behave like the default merge.data.frame function does, but introducing this w/o warning can possibly break code that uses this already.

If that sounds reasonable, maybe the default can be changed in data.table v1.6 (or something), otherwise that note in the documentation can be removed.

Add test 255.



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/datatable@199 edb9625f-4e0d-4859-8d74-9fd3b1da38cb